pull out ch-frb-l1::main() into ch-frb-l1-main.cpp
